implemented refresh_keys method
this method should be used to "refresh" the key/prev_key objects.

Currently it is called when in the constructor, after uploading but
before archiving, and after promoting.

This logic is slightly more involved and will need a bit more
baby sitting but it is more optimized, it only goes out to S3
when we ask it to, not everytime self.key or self.prev_key is
accessed.

	modified:   s3p/pipeline.py
	modified:   s3p/release.py